requirements.txt
streamlit
pandas
openai
PyPDF2
google-cloud-storage
google-cloud-secret-manager
# On déclare les paquets séparément pour éviter le bug de résolution
psycopg2-binary==2.9.9
google-cloud-kms
requests==2.32.3
firebase-admin
requests
streamlit-oauth
google-cloud-firestore
stripe